projects/xhs-image-filter/requirements.txt
# 小红书素材整理自动化工具依赖

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0

# 人脸检测
mediapipe>=0.10.0

# CLI界面
click>=8.0.0
rich>=13.0.0

# 其他工具
numpy>=1.24.0
